enter dev container as non-root user
Logs into the dev container as the non-root "tuas" user. This is to
prevent being root inside the container and modifying file permissions
on the host system.

Also moved Dockerfile to .devcontainer folder and checked in
devcontainer.json to git. To have the terminal use bash as it's shell,
the json file needed to be modified and now we have the same
configuration across all users. There's also a popup that shows up to
launch the container since VS Code will detect the devcontainer.json
file that's already there. https://stackoverflow.com/questions/55957783/should-i-check-in-the-vs-code-dev-container-files-for-example-the-devcontaine