move Dockerfile to root of the project
We incorrectly assumed that you needed separate Dockerfiles for
devcontainers and just running an application. VS Code will ignore
whatever files you COPY into the container and just mount the files of
the project at /workspaces/project-name. It also doesn't seem to run
whatever command is specified with CMD or ENTRYPOINT.

Now the devcontainer.json file points to the Dockerfile at the root of
the repository.

Also, when we have more Dockerfiles for ARM machines or the Jetson, we
can have them all be in the root of the project.