Lint the code
In C++, the [[nodiscard]] attribute is used to indicate that the return value of a function should not be ignored. It serves as a reminder to the developer to handle the return value appropriately. When this attribute is applied to a function, if its return value is discarded without being used or assigned, the compiler will generate a warning or an error. This helps catch potential bugs where a function's return value is unintentionally ignored, ensuring that it is properly handled.

{} after a member variable zero initializes it.
https://stackoverflow.com/questions/61520931/what-do-curly-braces-after-a-struct-variable-member-mean

const after a method:
"const means that the method promises not to alter any members of the class"
https://stackoverflow.com/questions/751681/what-is-the-meaning-of-const-at-the-end-of-a-member-function-declaration