mock camera implementation
implements two methods of CameraInterface to quickly plug in something
to the rest of the pipeline. For now it just generates an image of a
solid color. In the future we could have it return an actual competition
image with a valid target that the models can recognize.

set up other CameraInterface related functions to get MockCamera working