temporarily remove arena_sdk download
The wget command that pulls arena_sdk from our google drive is broken
and is failing with a virus check error. I can't figure out why it's not
working so I tried to switch to gdown. Then I was having issues with
running gdown since it's installed under the "tuas" user in the
Dockerfile and for some reason the github action is running inside the
container as root. I tried to pass some options into the runner to
change the container user but it was not getting set somehow. So, this
is the best solution to keep the unit tests running on github. It's not
too big of a deal right now since those that need arena sdk can manually
download it.

I am working to get this resolved by installing it in the Docker image https://github.com/tritonuas/obcpp/issues/67