BREAKING: CMake Surrenders!
At 9:06PM on April 6th, 2024 CMake officially surrendered to the
allied TUAS forces after suffering a massive defeat on the Western
Front. CMake had been steadily losing ground for hours, but was
still holding on due to the obscure and arcane differences between
STATIC libraries and SHARED libraries. Peace terms are expected to be
negotiated throughout the coming days.