The Great CMake Library War (#139)
* change output to DetectedTarget vector with match distance exposed

* start cv aggregator

* start worker logic

* finish base of cvaggregator

* add cv_aggregator integration test

* integrate cv aggregator with mission state

* fix lint

* code review feedback + general cleanup

* remove experimental cmake change (to be added later :eyes:)

* working for main binary

* before i completly fuck everything up

* obcpp compiling but not actually working

* BREAKING: CMake Surrenders!

At 9:06PM on April 6th, 2024 CMake officially surrendered to the
allied TUAS forces after suffering a massive defeat on the Western
Front. CMake had been steadily losing ground for hours, but was
still holding on due to the obscure and arcane differences between
STATIC libraries and SHARED libraries. Peace terms are expected to be
negotiated throughout the coming days.

* pass lint

* switch to ninja for ~33% faster compile times

* switch back to make temporarily