move aggregator CVResults index to std::optional
Originally the indices to the detected target vector were stored as size_t.
On the devcontainer size_t is mapped to unsigned long and the indices
were getting set to -1 in some places which was causing a < 0 check to
fail. Instead, an optional should be more expressive in expressing if a
bottle has been matched or not.